+++ pkgs/html/lib/src/token.dart
@@ -1,3 +1,7 @@
+// Copyright (c) project authors. All rights reserved.
+// Licensed under the MIT license.
+// See LICENSE file in the project root for details.
+
 /// This library contains token types used by the html5 tokenizer.
 library;