+++ pkgs/yaml/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.1.4-wip
+
+* Fix a stack overflow in `deepHashCode` when handling self-referential lists.
+
 ## 3.1.3
 
 * Require Dart 3.4